Don't break reporting if attachment file is wrong
